requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
pytest==8.4.1
alembic==1.16.4
python-dotenv==1.1.1
